+++ src/main/groovy/com/bmuschko/gradle/cargo/util/FileUtil.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.bmuschko.gradle.cargo
+package com.bmuschko.gradle.cargo.util
 
-import groovy.util.logging.Slf4j
-
-/**
- * Type of deployable.
- *
- * @author Benjamin Muschko
- */
-@Slf4j
-enum DeployableType {
-    WAR('war'), EAR('ear'), EXPLODED('war')
-
-    final String type
-
-    private DeployableType(String type) {
-        this.type = type
-    }
+interface FileUtil {
+    String getExtension(File file)
 }